Move license default behavior into library
Moves the handling of `['newrelic']['license']`, `['newrelic']['server_monitoring']['license']`, and `['newrelic']['application_monitoring']['license']` into a library. This defers the calculation of which value to use until the actual runtime, allowing wrapper cookbooks to set `['newrelic']['license']` and still have the value they set be usable for the other two.

This solves the complaint in https://github.com/escapestudios-cookbooks/newrelic/issues/101.