Ensure setting execute_php5enmod is meaningful
Prior to this, two instances of a variable assignment like so:

execute_php5enmod = new_resource.execute_php5enmod ? 'true' : 'false'

would end up meaning that `execute_php5enmod` was always truthy.

I.e., `'false' == true` in Ruby

As the value is meant to be a boolean no matter what, and in the
included recipe is cast to a boolean value, we should rely on its
boolean-ness instead of always being true.