Added support for two stage configuration of java agent
This patch fix all the java_agent integration tests for centos, fix also
the unit test for the previous patch sent by Juan Carlos and add support
for two stage configuration of java agent.

We are using chef to bake amazon images, during that bake process we
have access to any external repository but when we launch that AMI in
the VPC we do not have access to any external repository, that is why we
need this feature. It is not chaging the current behaviour of the
cookbook, but adding an extra feature.